Added specific version command to the command factory and fix errors
