+++ BotFlip.vcxproj.filters
@@ -66,7 +66,10 @@
     <ClCompile Include="ModCommand.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="VersionCommand.cpp">
+    <ClCompile Include="SpcfcVer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="VerCommand.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
